requirements.txt
flask>=2.0.0
tensorflow>=2.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.0.0
gdown>=4.0.0
waitress>=2.0.0
streamlit>=1.28.0
